share getTextLabelElm
